Fixed 2 bugs occured in 2D and Columbus mode
Bug fixed: determining center did not work when in 2D or Columbus mode. This can be achieved by using camera.pickEllipsoid.

Bug fixed: If no center could be found (e.g. when in Columbus mode and the mid-ray does not intersect the map) then Camera.flyTo is used. Previously Camera.zoomTo was used but with wrong arguments which resulted in a DeveloperError.